JWT Middleware updated with isStudent and isInstructor functions.
